Adds JSON configurations file to simplify configuration.
This adds a configurations.json file, which is set to be copied to the output directory through a build event.
This file has a connectionString property to determine the database connection, and a qtyRows option which determines the max number of rows to affect with the processing.
As a major change, modifies the DBConnection base class to accept a connectionString argument which should be passed on instantiation.
As a major change, now the methods to process the databasee are not static anymore. In order to instantiate the Updater object a connectionString argument must be provided. This is ment to allow the injection of different databases in case it's needed (you can just write the connection string on the initiation instead of reading the json file).
As a minor change, this reduces the arguments to pass to the metods fixing the limit to 1 record, and providing only the starting point for that iteration.
This also adds a new method to test the DB Connection called TestReadDBAsync which will read the first field of the table from start to end.